Se dividen las actions del flux (contexto global de la app) en partes más pequeñas (por página), se realiza limpieza de archivos obsoletos del proyecto, actualizado Layout.jsx (element, en vez de component)
