+++ src/front/js/components/menu_navbar.jsx
@@ -7,7 +7,7 @@ import youtube from "../../img/youtube.png";
 export const Menu_navbar = ({ setOpen }) => {
   const handleOptionClick = (option) => {
     if (option === "close") {
-      setOpen(false); // Cerrar el canvas cuando se hace clic en "X"
+      setOpen(false); // Se cierra el canvas cuando se hace clic en "X"
     }
   };
 
@@ -41,11 +41,6 @@ export const Menu_navbar = ({ setOpen }) => {
         <div className="py-8 flex justify-end text-6xl">
           <a className="hover:text-shape_red" href="/vlog">
             Vlogs
-          </a>
-        </div>
-        <div className="py-8 flex justify-end text-6xl">
-          <a className="hover:text-shape_red" href="#">
-            Tienda
           </a>
         </div>
         <div className="flex w-full mt-24 items-center justify-end">
@@ -91,5 +86,5 @@ export const Menu_navbar = ({ setOpen }) => {
   );
 };
 Menu_navbar.propTypes = {
-  setOpen: PropTypes.func.isRequired, // PropTypes para setOpen como función
+  setOpen: PropTypes.func.isRequired,
 };
