--- src/front/js/components/modal.jsx
@@ -1,31 +1,0 @@
-// // import { Login } from "../pages/login"
-// import PropTypes from 'prop-types';
-
-// export const Modal = ({ onClose , chefName}) => {
-
-//     Modal.propTypes = {
-//         onClose: PropTypes.any,
-//         chefName: PropTypes.string,
-//       };
-
-//     return (
-//         <div className="modal" tabIndex="-1" style={{ display: 'block' }}>
-//             <div className="modal-dialog">
-//                 <div className="modal-content">
-//                     <div className="modal-header">
-//                         <h5 className="modal-title">{chefName}</h5>
-//                         <button type="button" className="btn-close" onClick={onClose} aria-label="Close"></button>
-//                     </div>
-//                     <div className="modal-body">
-//                         <p>Modal body text goes here.</p>
-//                     </div>
-//                     <div className="modal-footer">
-//                         <button type="button" className="btn btn-secondary" onClick={onClose}>Close</button>
-//                     </div>
-//                 </div>
-//             </div>
-//         </div>
-//     );
-// };
-
-
