+++ src/front/js/pages/setting/mocks.js
@@ -56,34 +56,3 @@ export const changeCard = [
         className: 'block mt-2 mb-4 bg-shape_input w-full rounded-full border-0 py-1.5 pl-4 pr-72 text-gray-900 placeholder:text-gray-400 sm:text-sm sm:leading-6'
     },
 ]
-export const change = [
-    {
-      id: 1,
-      name: "Ingredientes para 1",
-      type: "textarea",
-      className:
-        "mt-2 mb-4 bg-shape_input w-96 h-72 rounded-xl border-0 py-1.5 pl-4 text-black placeholder:text-gray-500 sm:text-sm sm:leading-6 mr-36",
-      placeholder: `Ingredientes para masa: 
-          250 gramos de Harina de trigo
-          15 gramos de Levadura prensada
-          1 taza de Agua (240 ml)
-          1 pizca de Sal
-          1 cucharadita de Aceite`,
-    },
-    {
-      id: 2,
-      name: "Ingredientes para 2",
-      type: "textarea",
-      className:'mt-2 mb-4 bg-shape_input w-96 h-72 rounded-xl border-0 py-1.5 pl-4 text-black placeholder:text-gray-500 sm:text-sm sm:leading-6 mr-48',
-      placeholder: `Para la cubierta:
-        150 centímetros cúbicos de Salsa de tomate
-        150 gramos de Queso mozzarella
-        1 lata de Atún en aceite
-        100 gramos de Aceitunas de Origen
-        1 cucharadita de Orégano
-        2 cucharadas soperas de Queso rallado
-        1 pizca de Sal y Pimienta
-        9 Aceitunas (opcional)
-        50 centímetros cúbicos de Aceite de oliva`,
-    },
-  ];
