--- src/front/js/pages/vlog/mocks.js
@@ -1,27 +1,0 @@
-// src/data/videoData.js
-export const videos = [
-    {
-      id: 0,
-      src: "src/front/img/img_vlog/pizza.mp4",
-      likes: 726,
-      comments: 258,
-      description: "Pizza de Atún"
-    },
-    {
-      id: 1,
-      src: "src/front/img/img_vlog/video_vlog.mp4",
-      likes: 124,
-      comments: 59,
-      description: "Cooking Tutorial"
-    },
-    {
-      id: 2,
-      src: "src/front/img/img_vlog/video_vlog.mp4",
-      likes: 342,
-      comments: 91,
-      description: "Delicious Pizza"
-    }
-    // Añade más videos según sea necesario
-  ];
-  
-
